Introduce RegistryReader, RegistryWriter, and RegistryAccessor interfaces
These interfaces allow for more fine-tuned control over access to a 
registry.

The Container only needs to read from a registry, so it can make do with
a RegistryReader.

Applications may only want to allow full access to a registry during
an initialization period, during which they expose a registry through
a RegistryAccessor.

These interfaces also provide the opportunity to normalize input, such
as specifiers, in a class that proxies to an underlying Registry.